use next unplayed match hook
